# speed up validation by caching info for types