This also opens the way for speeding up the individual validators,
which can now return a specialized validator function for a given
type.

I've added one benchmark to get some idea of the speedup
(the validated type is similar to one in some production code).

        name        old time/op  new time/op  delta
        Validate-4  10.6µs ± 2%   0.9µs ± 3%  -91.81%  (p=0.008 n=5+5)